file_explorer.rs doc tests passing on windows
Cleaning up failures when running `cargo test` on a fresh clone on
a Windows machine.

The previous doc tests for the `file_type` function used the unix
FileTypeExt to demonstrate testing to see if a file is a directory.
Which fails on windows. The tests were marked as `no_run`, but this
only prevents the doc test from running - `cargo test` will still try to
build the code.

Modified the documentation to use `FileType::is_dir`, which has been
stable since 2015 via Rust 1.1
[link](https://doc.rust-lang.org/std/fs/struct.FileType.html#method.is_dir).

Alternative implementations:
* use `ignore` instead of `no_run` which will actually cause doc test to
  fully ignore the sample code
* document something else a user could do with the file type