doc tests from README.md in lib.rs passing
Cleaning up failures when running `cargo test` on a fresh clone.

The sample code in `README.md` is pulled in as rust documentation
through an `include_str!` macro, but caused a doc test failure when
trying to access modules in the `crossterm` crate.

Similar to the code in `examples/`, modified the sample code in
`README.md` to access crossterm functionality through
`ratatui::crossterm` rather than directly through the `crossterm` crate.