set_cwd does not change display dir on failure
Previously, calling `FileExplorer::set_cwd` on a path that does not
exist (or fails the inner `get_and_set_files` call for any other reason)
would result in the FileExplorer claiming to be set in the erroneous
directory, while still listing the contents of whatever directory it was
previously set up from.

This can happen if a user manually handles the result (without
`unwrap`-ing or panicking, etc.), but continues to use the
`FileExplorer`.